% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dashboard.R
\name{build_dashboard}
\alias{build_dashboard}
\title{Build the `bdmonitor` Dashboard}
\usage{
build_dashboard()
}
\value{
Prints R Markdown dashboard in HTML
}
\description{
`build_dashboard()` is a function for generating the bdmonitor
dashboard. It is a flexdashboard that uses different R Markdown files as
dashboard pages.
}
\examples{
build_dashboard()

}
