% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dashboard.R
\docType{data}
\name{dashboard_name}
\alias{dashboard_name}
\title{To build CI flexdashboard}
\format{
An object of class \code{character} of length 1.
}
\usage{
dashboard_name
}
\value{
prints rmarkdown report in HTML
}
\description{
To generate CI dashboard
}
\keyword{datasets}
